requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
typer==0.21.1
click==8.3.1  # pin: Typer 0.9 compatibility; avoid make_metavar signature mismatch
jsonschema==4.26.0
rich==14.2.0
textual==7.0.1  # modern terminal UI framework
pyyaml>=6.0  # required for template system
ttkbootstrap>=1.10.1  # modern theming for Tk desktop UI
pytest==9.0.2
openai>=1.30.0  # optional for examples/openai_pipe.py
python-dotenv>=1.0.0  # required for loading .env (Groq API)
httpx<0.29  # pin: Starlette TestClient older style requires httpx<0.29
h11>=0.16.0 # not directly required, pinned by Snyk to avoid a vulnerability
anyio>=4.4.0 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
pillow>=10.3.0 # not directly required, pinned by Snyk to avoid a vulnerability

Jinja2>=3.1.0  # required for report generation
tiktoken>=0.7.0  # required for token counting
